+++ isitfit/cost/mainManager.py
@@ -4,7 +4,6 @@ from tqdm import tqdm
 import datetime as dt
 import numpy as np
 import pytz
-from .cacheManager import RedisPandas as RedisPandasCacheManager
 from .datadogManager import HostNotFoundInDdg, DataNotFoundForHostInDdg
 
 import logging
@@ -34,6 +33,7 @@ def tagsContain(f_tn, ec2_obj):
   return False
 
 
+from isitfit.cost.cacheManager import RedisPandas as RedisPandasCacheManager
 class MainManager:
     def __init__(self, ctx, ddg=None, filter_tags=None):
         # set start/end dates
@@ -250,31 +250,7 @@ And finally re-run isitfit as usual.
           return None
 
         # check cache first
-        cache_key = "mainManager._get_ddg_cached/%s"%ec2_obj.instance_id
-        if self.cache_man.isReady():
-          df_cache = self.cache_man.get(cache_key)
-          if df_cache is not None:
-            logger.debug("Found datadog metrics in redis cache for %s, and data.shape[0] = %i"%(ec2_obj.instance_id, df_cache.shape[0]))
-            return myreturn(df_cache)
-
-        # if no cache, then download
-        df_fresh = pd.DataFrame() # use an empty dataframe in order to distinguish when getting from cache if not available in cache or data not found but set in cache
-        try:
-          df_fresh = self.ddg.get_metrics_all(ec2_obj.instance_id)
-        except HostNotFoundInDdg:
-          pass
-        except DataNotFoundForHostInDdg:
-          pass
-
-        # if caching enabled, store it for later fetching
-        # https://stackoverflow.com/a/57986261/4126114
-        # Note that this even stores the result if it was "None" (meaning that no data was found)
-        if self.cache_man.isReady():
-          # print("Saving to redis %s"%ec2_obj.instance_id)
-          self.cache_man.set(cache_key, df_fresh)
-
-        # done
-        return myreturn(df_fresh)
+        return self.ddg.get_metrics_all(host_id)
 
 
     def _handle_ec2obj(self, ec2_obj):